pyproject.toml
[tool.poetry]
name = "simple-gdrive"
version = "1.0.0"
description = "Simple command to help download and upload big files to Google Drive (using their official Drive APIs)"
authors = ["Binh Vu <binh@toan2.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.47.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
click = "^8.1.3"
tqdm = "^4.64.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
